requirements.txt
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
fonttools==4.25.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numpy==1.20.3
olefile==0.46
Pillow==8.3.1
pip==21.0.1
psutil==5.8.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
pyspng==0.1.0
python-dateutil==2.8.2
python-rclone==0.0.2
requests==2.26.0
scipy==1.6.2
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.16.0
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tornado==6.1
tqdm==4.62.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wheel==0.37.0
zipp==3.5.0
